# Conflicts:
#	Zimply/app/src/main/java/com/application/zimplyshop/activities/NewProductDetailActivity.java
#	Zimply/app/src/main/java/com/application/zimplyshop/activities/RecentProductsActivity.java
#	Zimply/app/src/main/java/com/application/zimplyshop/activities/SearchResultsActivity.java
#	Zimply/app/src/main/java/com/application/zimplyshop/adapters/ProductsRecyclerViewGridAdapter.java
#	Zimply/app/src/main/java/com/application/zimplyshop/fragments/ProductSearchFragment.java
#	Zimply/app/src/main/java/com/application/zimplyshop/utils/ZTracker.java

+++ Zimply/app/src/main/java/com/application/zimplyshop/objects/AllProducts.java
@@ -480,7 +480,7 @@ public class AllProducts {
         this.category_tree = category_tree;
     }
 
-    public boolean cartContains(int productId){
+    public boolean cartContains(long productId){
         if(cartObjs!=null && cartObjs.size()>0){
             for(BaseCartProdutQtyObj obj: cartObjs){
                 if(obj.getId() == productId){
